move conencted_peers out of Store
simplify Store trait, don't report conencted unless confirmed by swarm event
